schedular/.history/requirements_20241107230144.txt
# Core Django dependencies
Django==4.2  
djangorestframework==3.14  
pytz==2023.3  # For time zone management
django-timezone-field==5.0  # For user time zone preferences
django-celery-beat==2.5.0  # For periodic tasks and scheduling
celery==5.3.0 
psycopg2-binary==2.9.6  # PostgreSQL
python-dotenv==1.0.0  # For managing environment variables
drf-yasg==1.21.6  # For auto-generating API documentation (Swagger)
djangorestframework-simplejwt==5.2.2  # JWT authentication for DRF